Add files via upload
EmailParser -- A method to increase speed of consuming email content

DESCRIPTION
--------------
This is a python script that is meant to be used with a text file export from Microsoft Outlook.  The result of running this python script is a textfile labled as 'Reading File for [Date].'  As an example, 'Reading File for March 31.'  This file  presents only email content from the chosen day, with the email chain truncated to only emails that were sent on the chosen day.  Duplictes are removed.